# src/main/kotlin/com/ivansuvorov/secretstash/api/SecretNoteController.kt
package com.ivansuvorov.secretstash.api

import com.ivansuvorov.secretstash.api.auth.JwtAuthInterceptor.Companion.USER_REQUEST_ATTRIBUTE
import com.ivansuvorov.secretstash.api.model.SecretNote
import com.ivansuvorov.secretstash.api.model.SecretNoteCreateRequest
import com.ivansuvorov.secretstash.api.model.SecretNotePublicLink
import com.ivansuvorov.secretstash.api.model.SecretNoteUpdateRequest
import com.ivansuvorov.secretstash.configuration.properties.PublicProperties
import com.ivansuvorov.secretstash.service.RateLimiterService
import com.ivansuvorov.secretstash.service.SecretNoteService
import com.ivansuvorov.secretstash.service.model.SecretNoteCreateRequestDto
import com.ivansuvorov.secretstash.service.model.SecretNoteDto
import com.ivansuvorov.secretstash.service.model.SecretNoteUpdateRequestDto
import com.ivansuvorov.secretstash.service.model.UserDto
import org.slf4j.MDC
import org.springframework.http.HttpStatus
import org.springframework.web.bind.annotation.DeleteMapping
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.PutMapping
import org.springframework.web.bind.annotation.RequestAttribute
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.ResponseStatus
import org.springframework.web.bind.annotation.RestController
import org.springframework.web.server.ResponseStatusException
import java.util.UUID

@RestController
@RequestMapping("/notes")
class SecretNoteController(
    private val secretNoteService: SecretNoteService,
    private val rateLimiterService: RateLimiterService,
    private val publicProperties: PublicProperties
) {
    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    fun createSecretNote(
        @RequestBody createRequest: SecretNoteCreateRequest,
        @RequestAttribute(USER_REQUEST_ATTRIBUTE) user: UserDto,
    ): SecretNote {
        rateLimiterService.checkForUser(user.id)

        MDC.put("requestId", UUID.randomUUID().toString())
        MDC.put("userId", user.id.toString())

        val secretNote =
            secretNoteService.create(
                caller = user,
                request = SecretNoteCreateRequestDto(
                    title = createRequest.title,
                    content = createRequest.content,
                    expiresAt = createRequest.expiresAt,
                ),
            )
        return secretNote.toApiModel()
    }

    @GetMapping("/{id}")
    fun getSecretNote(
        @PathVariable id: UUID,
        @RequestAttribute(USER_REQUEST_ATTRIBUTE) user: UserDto,
    ): SecretNote? {
        rateLimiterService.checkForUser(user.id)

        MDC.put("requestId", UUID.randomUUID().toString())
        MDC.put("userId", user.id.toString())
        MDC.put("secretNoteId", id.toString())

        val secretNote =
            secretNoteService.findById(
                caller = user,
                noteId = id,
            )
        return secretNote?.toApiModel()
    }

    @GetMapping
    fun getLatestSecretNotes(
        @RequestParam("count", required = false, defaultValue = "1000") count: Int,
        @RequestAttribute(USER_REQUEST_ATTRIBUTE) user: UserDto,
    ): List<SecretNote> {
        rateLimiterService.checkForUser(user.id)

        MDC.put("requestId", UUID.randomUUID().toString())
        MDC.put("userId", user.id.toString())

        if (count > 1000) {
            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "Can't retrieve more than 1000 notes")
        }

        val secretNotes =
            secretNoteService.findLatest(
                caller = user,
                count = count,
            )
        return secretNotes.map { it.toApiModel() }
    }

    @PutMapping("/{id}")
    fun updateSecretNote(
        @PathVariable id: UUID,
        @RequestBody updateRequest: SecretNoteUpdateRequest,
        @RequestAttribute(USER_REQUEST_ATTRIBUTE) user: UserDto,
    ): SecretNote {
        rateLimiterService.checkForUser(user.id)

        MDC.put("requestId", UUID.randomUUID().toString())
        MDC.put("userId", user.id.toString())
        MDC.put("secretNoteId", id.toString())

        val secretNote =
            secretNoteService.update(
                caller = user,
                noteId = id,
                request =
                    SecretNoteUpdateRequestDto(
                        title = updateRequest.title,
                        content = updateRequest.content,
                        expiresAt = updateRequest.expiresAt,
                    ),
            )
        return secretNote.toApiModel()
    }

    @DeleteMapping("/{id}")
    fun deleteSecretNote(
        @PathVariable id: UUID,
        @RequestAttribute(USER_REQUEST_ATTRIBUTE) user: UserDto,
    ) {
        rateLimiterService.checkForUser(user.id)

        MDC.put("requestId", UUID.randomUUID().toString())
        MDC.put("userId", user.id.toString())
        MDC.put("secretNoteId", id.toString())

        secretNoteService.delete(
            caller = user,
            noteId = id,
        )
    }

    @PostMapping("/{id}/share")
    fun shareSecretNote(
        @PathVariable id: UUID,
        @RequestAttribute(USER_REQUEST_ATTRIBUTE) user: UserDto,
    ): SecretNotePublicLink {
        rateLimiterService.checkForUser(user.id)

        MDC.put("requestId", UUID.randomUUID().toString())
        MDC.put("userId", user.id.toString())

        secretNoteService.shareNote(user, id)

        return SecretNotePublicLink(
            link = "${publicProperties.host}/public/note/$id"
        )
    }

    private fun SecretNoteDto.toApiModel(): SecretNote = SecretNote(
        id = id,
        title = title,
        content = content,
        expiresAt = expiresAt,
        createdAt = createdAt,
    )
}